requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==7.1.0
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
exceptiongroup==1.1.0
flake8==6.0.0
iniconfig==2.0.0
kombu==5.2.4
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
openpyxl==3.1.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dotenv==0.21.1
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
vine==5.0.0
wcwidth==0.2.6
